tweak wording and remove obtrusive CTA from cody chat messages
A "get Cody in your editor" CTA was showing in the bottom right of every single chat message. This was obtrusive. There is already a CTA at the top of the page to get Cody in your editor.

Also generally prefers "editor" to "IDE" except in the specific context of JetBrains IDEs.